refactor hles transformation
