+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/common/RawRecord.scala
@@ -1,8 +1,8 @@
 package org.broadinstitute.monster.dap.common
 
-import org.broadinstitute.monster.dap.HLESurveyTransformationPipelineBuilder.logger
-import org.broadinstitute.monster.dap.TruncatedDecimalError
+import org.broadinstitute.monster.dap.hles.HLESurveyTransformationPipelineBuilder.logger
 import org.broadinstitute.monster.dap.common.RawRecord.DAPDateTimeFormatter
+import org.broadinstitute.monster.dap.hles.TruncatedDecimalError
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
