+++ dap-etl/transformation/src/main/scala/org/broadinstitute/monster/dap/hles/HLESurveyTransformationLog.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.hles
 
 import com.spotify.scio.ScioMetrics.counter
 import org.broadinstitute.monster.dap.common.PostProcess.errCount
