+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/HLESurveyTransformationPipelineValidationSpec.scala
@@ -16,6 +16,7 @@ import org.broadinstitute.monster.dap.common.{
   HLESurveyTransformationFailException,
   PostProcess
 }
+import org.broadinstitute.monster.dap.hles.HLESurveyTransformationPipelineBuilder
 import upack.Msg
 
 import scala.io.Source
