+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/dog/DemographicsTransformationsSpec.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.monster.dap.dog
 
-import org.broadinstitute.monster.dap.MissingCalcFieldError
 import org.broadinstitute.monster.dap.common.RawRecord
+import org.broadinstitute.monster.dap.hles.MissingCalcFieldError
 import org.broadinstitute.monster.dogaging.jadeschema.fragment.HlesDogDemographics
 import org.scalatest.OptionValues
 import org.scalatest.flatspec.AnyFlatSpec
